backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
psycopg2-binary==2.9.3
python-dotenv==0.19.2
bcrypt==4.0.1
pandas==1.3.5
openpyxl==3.0.10